+++ alynt-plugin-updater.php
@@ -3,7 +3,7 @@
  * Plugin Name:       Alynt Plugin Updater
  * Plugin URI:        https://github.com/NichlasB/alynt-plugin-updater
  * Description:       Enable automatic updates for GitHub-hosted WordPress plugins.
- * Version:           1.1.2
+ * Version:           1.1.3
  * Requires at least: 6.0
  * Requires PHP:      7.4
  * Author:            Alynt
@@ -21,7 +21,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 	exit;
 }
 
-define( 'ALYNT_PU_VERSION', '1.1.2' );
+define( 'ALYNT_PU_VERSION', '1.1.3' );
 define( 'ALYNT_PU_PLUGIN_FILE', __FILE__ );
 define( 'ALYNT_PU_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'ALYNT_PU_PLUGIN_URL', plugin_dir_url( __FILE__ ) );